# openapi.yaml
openapi: 3.1.0
info:
  title: Sleeper Backend API
  description: |
    A comprehensive fantasy football API that aggregates data from **KeepTradeCut (KTC)** and **Sleeper API**.

    This API provides comprehensive fantasy football data by combining:
    - 🏈 **KeepTradeCut (KTC)**: Player rankings and trade values (Dynasty & Redraft rankings, 1QB & Superflex league formats, TEP scoring variations)
    - 🏈 **Sleeper API**: Player data & league management (Comprehensive player profiles, League management, Research data and projections)

    ## Performance Optimizations
    - **Database Caching Strategy**: First API call fetches from external APIs and caches in database, subsequent calls served instantly from local cache
    - **Response Time Guidelines**: First refresh call takes 30-60 seconds, cached data retrieval < 1 second, health checks < 100ms
    - **Gunicorn Multi-Worker Setup**: 9 worker processes for concurrent request handling with load balancing
  version: 1.0.0
  contact:
    name: API Support
    url: https://github.com/rawarne2/sleeper-backend
servers:
  - url: http://localhost:5001
    description: Development server
  - url: https://sleeper-backend.vercel.app
    description: Production server
...
paths:
...
  /api/dashboard/league/{league_id}:
    get:
      tags:
        - Dashboard
      summary: Get dashboard league bundle
      description: |
        Returns a merged dashboard payload for a league, including rosters, player projections/values,
        research metadata, and optional per-player season stats.
      operationId: getDashboardLeagueBundle
      parameters:
        - $ref: "#/components/parameters/LeagueId"
        - $ref: "#/components/parameters/SeasonQuery"
        - $ref: "#/components/parameters/IsRedraft"
        - $ref: "#/components/parameters/LeagueFormat"
        - $ref: "#/components/parameters/TepLevel"
      responses:
        "200":
          description: Dashboard payload retrieved successfully
          content:
            application/json:
              schema:
                type: object
        "400":
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: League not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/maintenance/nightly-sync:
    get:
      tags:
        - Maintenance
      summary: Run nightly sync pipeline (cron style)
      description: |
        Runs the ingest pipeline (KTC -> leagues -> research). Requires
        Authorization: Bearer <CRON_SECRET> in production.
      operationId: runNightlySyncGet
      responses:
        "200":
          description: Nightly sync completed
          content:
            application/json:
              schema:
                type: object
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags:
        - Maintenance
      summary: Run nightly sync pipeline (manual POST)
      description: |
        Same pipeline as GET /api/maintenance/nightly-sync with optional JSON body controls
        (league_ids, seasons, research_week, skip_ktc, skip_leagues, skip_research, skip_prewarm).
      operationId: runNightlySyncPost
      responses:
        "200":
          description: Nightly sync completed
          content:
            application/json:
              schema:
                type: object
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/maintenance/prewarm:
    get:
      tags:
        - Maintenance
      summary: Prewarm dashboard caches
      description: Prewarms known dashboard league cache keys; requires cron authorization.
      operationId: prewarmDashboardCaches
      responses:
        "200":
          description: Prewarm completed successfully
          content:
            application/json:
              schema:
                type: object
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: One or more prewarm requests failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags:
        - Sleeper Research
      summary: Refresh research data
      description: Manually refresh research data from Sleeper API for a specific season
      operationId: refreshResearchData
      parameters:
        - $ref: "#/components/parameters/Season"
        - $ref: "#/components/parameters/WeekQuery"
        - $ref: "#/components/parameters/LeagueType"
      responses:
        "200":
          description: Research data refreshed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchRefreshResponse"
        "400":
          description: Invalid parameters or failed to refresh
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      tags:
        - Sleeper Research
      summary: Update research data
      description: Manually refresh research data from Sleeper API for a specific season
      operationId: updateResearchData
      parameters:
        - $ref: "#/components/parameters/Season"
        - $ref: "#/components/parameters/WeekQuery"
        - $ref: "#/components/parameters/LeagueType"
      responses:
        "200":
          description: Research data refreshed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchRefreshResponse"
        "400":
          description: Invalid parameters or failed to refresh
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  parameters:
    IsRedraft:
      name: is_redraft
      in: query
      description: Ranking type - 'true' for redraft (current season only), 'false'
        for dynasty (long-term player value)
      required: false
      schema:
        type: string
        enum: [ "true", "false" ]
        default: "false"
      example: "false"
    LeagueFormat:
      name: league_format
      in: query
      description: League format - '1qb' for standard leagues, 'superflex' for
        superflex leagues (can start 2 QBs)
      required: false
      schema:
        type: string
        enum: [ "1qb", "superflex" ]
        default: "1qb"
      example: "superflex"
    TepLevel:
      name: tep_level
      in: query
      description: |
        TEP (Tight End Premium) scoring level:
        - '' (empty): Standard scoring
        - 'tep': +0.5 points per TE reception
        - 'tepp': +1.0 points per TE reception
        - 'teppp': +1.5 points per TE reception
      required: false
      schema:
        type: string
        enum: [ "", "tep", "tepp", "teppp" ]
        default: ""
      example: "tep"
    LeagueId:
      name: league_id
      in: path
      description: The Sleeper league ID
      required: true
      schema:
        type: string
        pattern: "^[0-9]+$"
      example: "1210364682523656192"
    Season:
      name: season
      in: path
      description: The NFL season year
      required: true
      schema:
        type: string
        pattern: "^[0-9]{4}$"
      example: "2024"
    SeasonQuery:
      name: season
      in: query
      description: The NFL season year
      required: false
      schema:
        type: string
        pattern: "^[0-9]{4}$"
        default: "2024"
      example: "2024"
...
    WeekQuery:
      name: week
      in: query
      description: Week number (1-18), or "all" for research refresh/read endpoints
      required: false
      schema:
        oneOf:
          - type: integer
            minimum: 1
            maximum: 18
          - type: string
            enum: [ "all" ]
        default: 1
      example: 1
...
    LeagueType:
      name: league_type
      in: query
      description: League type - "dynasty" or "redraft"
      required: false
      schema:
        type: string
        enum: [ "dynasty", "redraft" ]
        default: "dynasty"
      example: "dynasty"
  schemas:
...
    ErrorResponse:
      type: object
      description: >
        Standard JSON error envelope. Optional fields depend on the route;
        clients should tolerate unknown keys.
      required:
        - status
        - error
        - timestamp
      additionalProperties: true
      properties:
        status:
          type: string
          enum: [ "error" ]
        error:
          type: string
          description: Error message
        details:
          type: string
          description: >
            Additional detail as text. When the source value is structured
            (e.g. upstream API error object), the server may JSON-encode it into this string.
        timestamp:
          type: string
          format: date-time
          description: Error timestamp (RFC 3339, UTC with Z)
        hint:
          type: string
          description: Operator hint (e.g. maintenance auth)
        suggestion:
          type: string
          description: Suggested remediation (e.g. which refresh to run)
        job_id:
          type: string
        league_id:
          type: string
        season:
          type: string
        week:
          oneOf:
            - type: integer
            - type: string
          description: Week number or the literal "all" for research refresh errors
        league_type:
          type: string
          enum: [ "dynasty", "redraft" ]
        parameters:
          type: object
          additionalProperties: true
          description: Request parameters associated with the error (e.g. KTC format flags)
        configuration:
          type: object
          additionalProperties: true
          description: Server-side configuration context (e.g. cleanup target)
        results:
          type: array
          description: Per-item outcomes (e.g. prewarm summary entries)
          items:
            type: object
            additionalProperties: true
        sleeper_players_fetched:
          type: integer
          description: Count from Sleeper merge failures
...
    ResearchData:
      type: object
      properties:
        id:
          type: integer
        season:
          type: string
        week:
          type: integer
        league_type:
          type: string
          enum: [ "dynasty", "redraft" ]
        player_id:
          type: string
        research_data:
          type: object
          description: Research metrics and analytics data
        last_updated:
          type: string
          format: date-time
...
    ResearchRefreshResponse:
      type: object
      properties:
        status:
          type: string
          enum: [ "success" ]
        message:
          type: string
          example: "Research data refreshed successfully"
        season:
          type: string
        week:
          type: integer
        league_type:
          type: string
          enum: [ "dynasty", "redraft" ]
        refresh_results:
          type: object
        data:
          oneOf:
            - type: array
              items:
                $ref: "#/components/schemas/ResearchData"
            - type: object
        source:
          type: string
          enum: [ "database", "sleeper_api" ]
        database_saved:
          type: boolean
        timestamp:
          type: string
          format: date-time
...